+++ cloudformation/appmesh/aws-appmesh-route.go
@@ -24,7 +24,7 @@ type Route struct {
 	MeshOwner string `json:"MeshOwner,omitempty"`
 
 	// RouteName AWS CloudFormation Property
-	// Required: true
+	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-appmesh-route.html#cfn-appmesh-route-routename
 	RouteName string `json:"RouteName,omitempty"`