+++ cloudformation/databrew/aws-databrew-job_output.go
@@ -18,6 +18,11 @@ type Job_Output struct {
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-databrew-job-output.html#cfn-databrew-job-output-format
 	Format string `json:"Format,omitempty"`
 
+	// FormatOptions AWS CloudFormation Property
+	// Required: false
+	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-databrew-job-output.html#cfn-databrew-job-output-formatoptions
+	FormatOptions *Job_OutputFormatOptions `json:"FormatOptions,omitempty"`
+
 	// Location AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-databrew-job-output.html#cfn-databrew-job-output-location