+++ cloudformation/ecs/aws-ecs-cluster.go
@@ -28,11 +28,6 @@ type Cluster struct {
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-ecs-cluster.html#cfn-ecs-cluster-clustersettings
 	ClusterSettings []Cluster_ClusterSettings `json:"ClusterSettings,omitempty"`
 
-	// Configuration AWS CloudFormation Property
-	// Required: false
-	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-ecs-cluster.html#cfn-ecs-cluster-configuration
-	Configuration *Cluster_ClusterConfiguration `json:"Configuration,omitempty"`
-
 	// DefaultCapacityProviderStrategy AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-ecs-cluster.html#cfn-ecs-cluster-defaultcapacityproviderstrategy